+++ packages/indexing/src/index.ts
@@ -1,0 +1,4 @@
+export { get, set } from "./ops.js";
+export { get as get_with_setter } from "./get.js";
+export { set as set_with_setter } from "./set.js";
+export type * from "./types.js";
