+++ packages/ndarray/index.ts
@@ -1,8 +1,9 @@
 import ndarray from "ndarray";
 // @ts-expect-error
 import ops from "ndarray-ops";
-import { get_with_setter, set_with_setter } from "@zarrita/core";
 
+import { get_with_setter, set_with_setter } from "@zarrita/indexing";
+import type { Indices, Projection, GetOptions, Slice, SetOptions } from "@zarrita/indexing";
 import type * as core from "@zarrita/core";
 import type { Async, Readable, Writeable } from "@zarrita/storage";
 
@@ -10,7 +11,7 @@ export const setter = {
 	prepare: ndarray,
 	set_scalar<D extends core.DataType>(
 		dest: ndarray.NdArray<core.TypedArray<D>>,
-		selection: (number | core.Indices)[],
+		selection: (number | Indices)[],
 		value: core.Scalar<D>,
 	) {
 		ops.assigns(view(dest, selection), value);
@@ -18,7 +19,7 @@ export const setter = {
 	set_from_chunk<D extends core.DataType>(
 		dest: ndarray.NdArray<core.TypedArray<D>>,
 		src: ndarray.NdArray<core.TypedArray<D>>,
-		mapping: core.Projection[],
+		mapping: Projection[],
 	) {
 		const s = unzip_selections(mapping);
 		ops.assign(view(dest, s.to), view(src, s.from));
@@ -29,11 +30,11 @@ export const setter = {
 export async function get<
 	D extends core.DataType,
 	Store extends Readable | Async<Readable>,
-	Sel extends (null | core.Slice | number)[],
+	Sel extends (null | Slice | number)[],
 >(
 	arr: core.Array<D, Store>,
 	selection: Sel | null = null,
-	opts: core.GetOptions<Parameters<Store["get"]>[1]> = {},
+	opts: GetOptions<Parameters<Store["get"]>[1]> = {},
 ) {
 	return get_with_setter<D, Store, ndarray.NdArray<core.TypedArray<D>>, Sel>(
 		arr,
@@ -46,9 +47,9 @@ export async function get<
 /** @category Utility */
 export async function set<D extends core.DataType>(
 	arr: core.Array<D, (Readable & Writeable) | Async<Readable & Writeable>>,
-	selection: (null | core.Slice | number)[] | null,
+	selection: (null | Slice | number)[] | null,
 	value: core.Scalar<D> | ndarray.NdArray<core.TypedArray<D>>,
-	opts: core.SetOptions = {},
+	opts: SetOptions = {},
 ) {
 	return set_with_setter<D, ndarray.NdArray<core.TypedArray<D>>>(
 		arr,
@@ -60,8 +61,8 @@ export async function set<D extends core.DataType>(
 }
 
 function unzip_selections(
-	mapping: core.Projection[],
-): { to: (number | core.Indices)[]; from: (number | core.Indices)[] } {
+	mapping: Projection[],
+): { to: (number | Indices)[]; from: (number | Indices)[] } {
 	const to = [], from = [];
 	for (const m of mapping) {
 		if (m.to !== null) to.push(m.to);
@@ -73,7 +74,7 @@ function unzip_selections(
 /** Convert zarrita selection to ndarray view. */
 function view<D extends core.DataType>(
 	arr: ndarray.NdArray<core.TypedArray<D>>,
-	sel: (number | core.Indices)[],
+	sel: (number | Indices)[],
 ) {
 	const lo: number[] = [];
 	const hi: number[] = [];
