+++ packages/ndarray/package.json
@@ -16,6 +16,7 @@
 	"dependencies": {
 		"@types/ndarray": "^1.0.11",
 		"@zarrita/core": "workspace:^",
+		"@zarrita/indexing": "workspace:^",
 		"@zarrita/storage": "workspace:^",
 		"ndarray": "^1.0.19",
 		"ndarray-ops": "^1.2.2"
